Improve grass system: cull debug, terrain alignment, depth bias
Major grass subsystem improvements:

- Added DEV cull-debug mode with false-color visualization (reason codes for annulus, biome, frustum, near camera, bypass)
- Extracted terrain surface height calculation to shared TSL module for grass-terrain alignment
- Fixed grass-terrain z-fighting with depth bias and polygon offset constants
- Enhanced visibility tracking in compute: changed return from float to {visible, reason} for debug coloring
- Fixed compute queue disposal and async error handling
- Fixed frustum bypass detection and pitch-down cull edge cases
- Corrected index buffer type (Uint8Array → Uint16Array) and segment clamping (max 127)
- Fixed player glow additive blending (terrain props no longer dimmed by night multiplier)
- Updated flower/grass materials to support optional terrain surface position sampling

+++ src/world/grass/render/grassGeometry.ts
@@ -1,4 +1,4 @@
-// src/world/grass/grassGeometry.ts — tapered blade strip for SpriteNodeMaterial
+// src/world/grass/render/grassGeometry.ts — tapered blade strip for SpriteNodeMaterial
 import { BufferAttribute, BufferGeometry, StaticDrawUsage } from 'three';
 import { GRASS_CONFIG } from '../config/grassConfig';
 
@@ -26,7 +26,7 @@ export function createGrassBladeGeometry(options: GrassBladeGeometryOptions): Bu
 
   const positions = new Float32Array(vertexCount * 3);
   const uvs = new Float32Array(vertexCount * 2);
-  const indices = new Uint8Array(indexCount);
+  const indices = new Uint16Array(indexCount);
 
   const taper = (t: number) => halfWidthBase * (1 - 0.7 * t);
